% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{add_event_first_year}
\alias{add_event_first_year}
\title{extract a guess of the first year of even}
\usage{
add_event_first_year(x)
}
\arguments{
\item{x}{dwca list or table of events}
}
\value{
the input with eventYear added
}
\description{
extract a guess of the first year of even
}
