% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{event_first_year}
\alias{event_first_year}
\title{extract a guess of the first year of even}
\usage{
event_first_year(x)
}
\arguments{
\item{x}{dwca list or table of events}
}
\value{
numeric guess of the first year of an event
}
\description{
extract a guess of the first year of even
}
