% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_haedat}
\alias{fetch_haedat}
\title{Fetch HAEDAT data as DWCA from iOBIS or CSV from HADAET directly}
\usage{
fetch_haedat(where = c("obis", "iode")[1], ...)
}
\arguments{
\item{where}{char, one of 'obis' or 'haedat'}

\item{...}{other arguments}
}
\value{
0 for success, not-0 for failure
}
\description{
Fetch HAEDAT data as DWCA from iOBIS or CSV from HADAET directly
}
