% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_haedat_iode}
\alias{fetch_haedat_iode}
\title{Fetch HAEDAT iOBIS data as DWCA using the specified URL}
\usage{
fetch_haedat_iode(
  save_file = haedat_path(sprintf("\%s_haedat_search.csv", format(Sys.Date(),
    "\%Y-\%m-\%d"))),
  ...
)
}
\arguments{
\item{save_file}{character, the filename to download to}

\item{...}{other arguments for \code{\link[utils]{download.file}}}
}
\value{
the output of \code{\link[utils]{download.file}}
}
\description{
Fetch HAEDAT iOBIS data as DWCA using the specified URL
}
