% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_haedat_obis}
\alias{fetch_haedat_obis}
\title{Fetch HAEDAT iOBIS data as DWCA using the specified URL}
\usage{
fetch_haedat_obis(
  version = "3.6",
  base_uri = "https://ipt.iobis.org/hab/archive.do",
  save_file = haedat_path(sprintf("dwca-haedat-v\%s.zip", version[1])),
  ...
)
}
\arguments{
\item{version}{character, version number}

\item{base_uri}{charcater, the base URI upon which to build query}

\item{save_file}{character, the filename to download to}

\item{...}{other arguments for \code{\link[utils]{download.file}}}
}
\value{
the output of \code{\link[utils]{download.file}}
}
\description{
Fetch HAEDAT iOBIS data as DWCA using the specified URL
}
