% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haedat-package.R
\docType{package}
\name{haedat-package}
\alias{haedat-package}
\alias{_PACKAGE}
\alias{haedat}
\title{haedat:}
\description{
Simple access to downloading and reading HAEDAT data in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BigelowLab/haedat}
}

}
