% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{haedat_iode_uri}
\alias{haedat_iode_uri}
\title{Hides the messy URI for the HAEDAT csv table behind a function}
\usage{
haedat_iode_uri()
}
\value{
url for CSV download
}
\description{
Hides the messy URI for the HAEDAT csv table behind a function
}
