% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{list_dwca}
\alias{list_dwca}
\title{List DWCA resources, in order of most recent to oldest}
\usage{
list_dwca(path = haedat_path())
}
\arguments{
\item{path}{char, the path to list}
}
\value{
character vector of full filenames ordered by verion
}
\description{
List DWCA resources, in order of most recent to oldest
}
