% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_haedat_obis}
\alias{read_haedat_obis}
\title{Read OBIS haedat data in DWCA form}
\usage{
read_haedat_obis(
  what = c("event", "occurrence", "extendedmeasurementorfact"),
  filename = list_obis()[1],
  form = c("tibble", "sf")[2]
)
}
\arguments{
\item{what}{character, one or more data elements to read}

\item{filename}{character, the name of the file to read}

\item{form}{charcater, one of 'tibble' or 'sf' specify the output class type. Only
applies to the 'event' data source.  If form is 'sf' then the records with missing
location information (no lat and/or lon) are purged. Also, all remaining locations
are cropped to \code{[-180, -90, 180, 90]}.}
}
\value{
a list or one or more tibbles or sf objects
}
\description{
Read OBIS haedat data in DWCA form
}
\seealso{
\href{https://ipt.iobis.org/hab/resource?r=haedat}{IPT OBIS}
}
