% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_haedat_original}
\alias{read_haedat_original}
\title{Read haedat data in original form}
\usage{
read_haedat_original(
  filename = haedat_path("haedat.csv"),
  form = c("tibble", "sf")[2]
)
}
\arguments{
\item{filename}{character, the name of the file to read}

\item{form}{charcater, one of 'tibble' or 'sf' specify the output class type}
}
\value{
a tibble or sf object
}
\description{
Read haedat data in original form
}
\seealso{
\href{http://haedat.iode.org/eventSearch.php?listAll=1&sortby=name}{HAEDAT IODE}
}
