% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_haedat}
\alias{fetch_haedat}
\title{Fetch HAEDAT data using the specified URL}
\usage{
fetch_haedat(
  x = "https://ipt.iobis.org/hab/archive.do?r=haedat&v=3.2",
  save_file = haedat_path("dwca-haedat-v3.2.zip"),
  ...
)
}
\arguments{
\item{x}{character, URL defining resource to download}

\item{save_file}{character, the filename to download to}

\item{...}{other arguments for \code{\link[utils]{download.file}}}
}
\value{
the output of \code{\link[utils]{download.file}}
}
\description{
Fetch HAEDAT data using the specified URL
}
