% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{list_haedat}
\alias{list_haedat}
\title{List files/directories in the haedat root directory}
\usage{
list_haedat(path = haedat_path(), ...)
}
\arguments{
\item{path}{char, the path to list}

\item{...}{arguments passed to \code{\link[base]{list.files}}}
}
\value{
character vector
}
\description{
This is a wrapper around \code{\link[base]{list.files}}
}
