% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_haedat}
\alias{read_haedat}
\title{Read haedat data in original or DWCA form}
\usage{
read_haedat(source = c("dwca", "original"), ...)
}
\arguments{
\item{source}{character, one of 'orginal' or 'dwca'}

\item{...}{other arguments passed to \code{\link{read_haedat_original}} or \code{\link{read_haedat_dwca}}}
}
\value{
see \code{\link{read_haedat_original}} or \code{\link{read_haedat_dwca}}
}
\description{
Read haedat data in original or DWCA form
}
