% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_haedat_iode}
\alias{read_haedat_iode}
\title{Read IODE haedat data in CSV form}
\usage{
read_haedat_iode(filename = list_iode()[1], form = c("tibble", "sf")[2])
}
\arguments{
\item{filename}{character, the name of the file to read}

\item{form}{charcater, one of 'tibble' or 'sf' specify the output class type}
}
\value{
a tibble or sf object
}
\description{
Read IODE haedat data in CSV form
}
\seealso{
\href{http://haedat.iode.org/eventSearch.php?listAll=1&sortby=name}{HAEDAT IODE}
}
